Add Sidekiq initializer to skip SimpleCov
This is going to be a requirement for us to keep test coverage at 100%. The initializer file for Sidekiq includes configuration for the client_side and the server_side of Sidekiq. The server side is being used when running the worker dyno on Heroku, but there isn't a way to test that locally since we aren't deploying to Heroku within our test suite.